Fix the __main__ entry point.
Because we were importing __main__ in __init__, we ended up with
a warning that __main__ was already imported when it was executed
as main. Moving the previous contents of __main__.py to main.py
so that it can then be import in __init__.py and __main__.py fixes
that.
