+++ src/fixate/__init__.py
@@ -30,6 +30,6 @@ from fixate._switching import (
     generate_relay_matrix_pin_list as generate_relay_matrix_pin_list,
 )
 
-from fixate.__main__ import run_main_program as run
+from fixate.main import run_main_program as run
 
 __version__ = "0.6.2"
